image_classification_service/requirements.txt
fastapi
uvicorn
numpy==1.24.3
tensorflow==2.13.1
python-multipart
Flask==2.0.1
opencv-python==4.5.3.56
Pillow==8.4.0

# Add other required libraries here
